add derivative draft
the problem her is that we may get the same error rate for different beta,
since the same queries are captured for different alpha-intervals.

we now have to filter the least beta for a given error rate, as this is
what we're interested in. only then a derivative will make any sense.